Initial version of A09 to work under OSX
The assembler assumes ints are 16-bit etc, which Xcode doesn't support. Replacing ints with int_16t and similar for longs gets the code to work. Jeff Tranter reports that the checksum for S-records is wrong, and that appears to be correct. That will be the next challenge when I have time, but Intelhex files are fine so my EEPROM programmer is happy for now.